feature/1637/mainPageFiltersSearch added skills selector component
