+++ components/ui/SpecializationsSelector/index.vue
@@ -1,0 +1,406 @@
+<template>
+  <div class="page__skills skills">
+    <div
+      v-for="key in specCount"
+      :key="key"
+      class="skills__block block"
+    >
+      <div class="block__skill-spec">
+        <div class="block__specialization specialization">
+          <base-dd
+            v-model="displaySpecIndex[key]"
+            class="specialization__dd"
+            type="gray"
+            :items="specializationsNames"
+            :placeholder="$t('settings.selectSpec')"
+            :mode="'small'"
+            rules="required"
+            :label="$t('settings.specialization')"
+            :hide-selected="hideSelectedSpecs"
+            @input="switchSkill($event, key)"
+          />
+          <div class="specialization__skills skills">
+            <base-dd
+              v-model="skillIndex[key]"
+              class="specialization__dd"
+              :type="specIndex[key] < 0 ? 'disabled' : 'gray'"
+              :disabled="specIndex[key] < 0"
+              :placeholder="$t('settings.selectSkills')"
+              :items="skillsNames[displaySpecIndex[key]]"
+              :mode="'small'"
+              rules="required"
+              :label="$t('settings.skillsInput')"
+              :hide-selected="hideSelectedSkills[key]"
+              @input="addSkillToBadge($event, key)"
+            />
+            <div
+              v-if="selectedSkills[key].length === 5"
+              class="skills__error"
+            >
+              {{ $t('ui.buttons.errors.manySkills') }}
+            </div>
+          </div>
+        </div>
+        <div class="block__skill skill">
+          <div
+            v-for="(item, i) in selectedSkills[key]"
+            :key="i"
+            class="skill__badge"
+          >
+            {{ item.name }}
+            <button
+              class="skill__remove"
+              @click="removeSkillToBadge(item, key)"
+            >
+              <img
+                src="~assets/img/ui/close_blue.svg"
+                alt="x"
+              >
+            </button>
+          </div>
+        </div>
+      </div>
+      <base-btn
+        :text="$t('settings.removeSpec')"
+        class="specialization__btn specialization__btn_remove"
+        @click="removeSpecialization(key)"
+      />
+    </div>
+    <base-btn
+      :text="$t('settings.addSpec')"
+      :disabled="specCount === 3"
+      class="skills__btn-add"
+      :class="specCount === 3 ? 'skills__btn-add_disabled' : ''"
+      @click="addSpecialization"
+    />
+    <div
+      v-if="specCount === 3"
+      class="skills__error"
+    >
+      {{ $t('ui.buttons.errors.manySpec') }}
+    </div>
+  </div>
+</template>
+
+<script>
+import { mapGetters } from 'vuex';
+
+export default {
+  name: 'SpecializationsSelector',
+  props: {
+    skills: {
+      type: Array,
+      default: null,
+    },
+  },
+  data() {
+    return {
+      specCount: 0,
+      specIndex: { 1: -1, 2: -1, 3: -1 }, // Выбранные специализации по id в filters
+      displaySpecIndex: { 1: -1, 2: -1, 3: -1 }, // Выбранные для отображения названия специализации
+      hideSelectedSkills: { 1: [], 2: [], 3: [] },
+      skillIndex: { 1: -1, 2: -1, 3: -1 },
+      selectedSkills: { 1: [], 2: [], 3: [] },
+    };
+  },
+  computed: {
+    ...mapGetters({
+      filters: 'quests/getFilters',
+    }),
+    specializationsNames() {
+      const specs = [];
+      const keys = Object.keys(this.filters);
+      for (let i = 0; i < keys.length; i += 1) {
+        const key = keys[i];
+        specs.push(this.$t(`filters.items.${this.filters[key].id}.title`));
+      }
+      return specs;
+    },
+    specsIndexes() { // id спеки из filters
+      const specsIndexes = [];
+      const keys = Object.keys(this.filters);
+      for (let i = 0; i < keys.length; i += 1) {
+        const key = keys[i];
+        specsIndexes.push(this.filters[key].id);
+      }
+      return specsIndexes;
+    },
+    skillsNames() {
+      const skillsData = {};
+      const keys = Object.keys(this.filters);
+      for (let i = 0; i < keys.length; i += 1) {
+        const key = keys[i];
+        const skillsBySpec = this.filters[key].skills;
+        const skillsIndexes = Object.values(skillsBySpec);
+        const skillsNames = [];
+        for (let j = 0; j < skillsIndexes.length; j += 1) {
+          const name = this.$t(`filters.items.${this.filters[key].id}.sub.${skillsIndexes[j]}`);
+          skillsNames.push(name);
+        }
+        skillsData[i] = skillsNames;
+      }
+      return skillsData;
+    },
+    skillsIndexes() {
+      const skillsIndexesData = {};
+      const keys = Object.keys(this.filters);
+      for (let i = 0; i < keys.length; i += 1) {
+        const key = keys[i];
+        const skillsBySpec = this.filters[key].skills;
+        skillsIndexesData[i] = Object.values((skillsBySpec)); // skills indexes from filters
+      }
+      return skillsIndexesData;
+    },
+    hideSelectedSpecs() {
+      return Object.values(this.displaySpecIndex);
+    },
+  },
+  async mounted() {
+    if (!Object.keys(this.filters).length) await this.$store.dispatch('quests/getFilters');
+    if (this.skills) await this.fillData();
+  },
+  methods: {
+    async fillData() {
+      const specKeys = {};
+      let key = 1;
+      // eslint-disable-next-line no-restricted-syntax
+      for (const item of this.skills) {
+        const [_spec, _skill] = item.path.split('.');
+        const spec = parseInt(_spec, 10);
+        const skill = parseInt(_skill, 10);
+        if (!Object.keys(specKeys).includes(_spec)) {
+          specKeys[spec] = key;
+          this.displaySpecIndex[key] = this.specsIndexes.indexOf(spec);
+          this.selectedSkills[key] = [];
+          this.specIndex[specKeys[spec]] = spec;
+          key += 1;
+        }
+        const displaySkillIndex = this.skillsIndexes[this.displaySpecIndex[specKeys[spec]]].indexOf(skill);
+        if (!this.hideSelectedSkills[specKeys[spec]]) this.hideSelectedSkills[specKeys[spec]] = [displaySkillIndex];
+        else this.hideSelectedSkills[specKeys[spec]].push(displaySkillIndex);
+        this.selectedSkills[specKeys[spec]].push({
+          index: skill,
+          name: this.skillsNames[this.displaySpecIndex[specKeys[spec]]][displaySkillIndex],
+        });
+      }
+      this.specCount = key - 1;
+      this.onChangeInputSpecs();
+    },
+    onChangeInputSpecs() {
+      const specAndSkills = [];
+      let key = 1;
+      // eslint-disable-next-line guard-for-in,no-restricted-syntax
+      for (const spec in this.specIndex) {
+        if (this.specIndex[spec] !== -1) {
+          for (let i = 0; i < this.selectedSkills[key].length; i += 1) {
+            specAndSkills.push(`${this.specIndex[spec]}.${this.selectedSkills[key][i].index}`);
+          }
+        }
+        key += 1;
+      }
+      this.$emit('changeSkills', specAndSkills);
+    },
+    addSkillToBadge(event, key) {
+      const index = this.skillsIndexes[this.displaySpecIndex[key]][event];
+      if (this.selectedSkills[key].length > 4
+        || this.selectedSkills[key].filter((obj) => obj.index === index).length) return;
+      this.selectedSkills[key].push({
+        index,
+        name: this.skillsNames[this.displaySpecIndex[key]][event],
+      });
+      this.hideSelectedSkills[key].push(event);
+      this.onChangeInputSpecs();
+    },
+    removeSkillToBadge(skill, key) {
+      let hideIndex = null;
+      for (let i = 0; i < this.skillsIndexes[key].length; i += 1) {
+        if (this.skillsIndexes[this.displaySpecIndex[key]][i] === skill.index) {
+          hideIndex = i;
+          break;
+        }
+      }
+      const numberInHide = this.hideSelectedSkills[key].indexOf(hideIndex);
+      if (numberInHide > -1) {
+        this.hideSelectedSkills[key].splice(numberInHide, 1);
+      }
+
+      const numberInArray = this.selectedSkills[key].indexOf(skill);
+      this.selectedSkills[key].splice(numberInArray, 1);
+      if (!this.selectedSkills[key].length) {
+        this.skillIndex[key] = -1;
+      }
+      this.onChangeInputSpecs();
+    },
+    switchSkill(event, key) {
+      this.specIndex[key] = this.specsIndexes[event];
+      this.skillIndex[key] = -1;
+      this.selectedSkills[key] = [];
+      this.hideSelectedSkills[key] = [];
+      this.onChangeInputSpecs();
+    },
+    addSpecialization() {
+      if (this.specCount <= 2) {
+        this.specCount += 1;
+      }
+      this.onChangeInputSpecs();
+    },
+    removeSpecialization(key) {
+      this.selectedSkills[key] = [];
+      this.specIndex[key] = -1;
+      this.displaySpecIndex[key] = -1;
+      this.skillIndex[key] = -1;
+      this.specCount -= 1;
+      this.onChangeInputSpecs();
+    },
+  },
+};
+</script>
+
+<style lang="scss" scoped>
+.page {
+  &__skills {
+    width: 100%;
+    .block {
+      display: flex;
+      grid-gap: 20px;
+      justify-content: space-between;
+      margin-top: 20px;
+      &__skill-spec {
+        width: 100%;
+      }
+      &__specialization {
+        display: flex;
+        align-items: flex-start;
+        flex-direction: row;
+        grid-gap: 20px;
+      }
+      &__skill {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        flex-wrap: wrap;
+        grid-gap: 10px;
+        .skill {
+          &__badge {
+            background: rgba(0, 131, 199, 0.1);
+            border-radius: 44px;
+            color: $blue;
+            white-space: nowrap;
+            grid-gap: 8px;
+            padding: 5px 10px 5px 10px;
+            display: flex;
+            text-align: center;
+            &-skills {
+              padding: 15px;
+            }
+          }
+        }
+      }
+    }
+  }
+  &__spec {
+    flex-direction: row;
+    flex-wrap: wrap;
+    display: flex;
+    max-width: 1180px;
+    width: 100%;
+    justify-content: flex-start;
+    //padding: 0 20px 0 0;
+  }
+  &__title {
+    @include text-simple;
+    margin: 30px 0 0 0;
+  }
+  &__raising {
+    @include text-simple;
+    font-weight: 500;
+    font-size: 20px;
+    color: $black800;
+    margin: 0 0 20px 0;
+  }
+  &__page {
+    font-weight: 500;
+    font-size: 25px;
+    color: $black800;
+  }
+  &__dd {
+    min-width: 160px;
+  }
+  &__category {
+    align-items: flex-start;
+    margin: 20px 0 0 0;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    grid-gap: 20px;
+  }
+  &__address {
+    margin: 20px 0 0 0;
+    display: grid;
+    grid-gap: 20px;
+  }
+  &__textarea {
+    @include text-simple;
+    border-radius: 6px;
+    padding: 11px 20px 11px 15px;
+    height: 214px;
+    width: 100%;
+    border: 0;
+    background-color: #F3F7FA;
+    resize: none;
+    &::placeholder {
+      color: $black300;
+    }
+  }
+}
+.specialization {
+  &__dd {
+    margin-bottom: 15px;
+    width: 100%;
+  }
+  &__skills {
+    width: 100%;
+  }
+  &__btn {
+    text-align: center;
+    &_remove {
+      margin-top: 37px;
+      width: 50%;
+      background: #ffffff;
+      color: #d73838;
+      border: 1px solid #e79a9a;
+    }
+    &_remove:hover {
+      background: #e79a9a;
+    }
+  }
+}
+.skills {
+  &__add-info {
+    display: flex;
+    flex-wrap: nowrap;
+    grid-gap: 20px;
+    margin-top: 20px;
+  }
+  &__btn {
+    &-add {
+      text-align: center;
+      margin-top: 20px;
+      width: 250px;
+      background: #FFFFFF;
+      color: #0083C7;
+      border: 1px solid #bce8ff;
+      &_disabled {
+        color: #aaaaaa !important;
+        border: 0.5px solid #aaaaaa;
+      }
+    }
+    &-add:hover {
+      background: #bce8ff;
+      }
+  }
+  &__error {
+    color: #f36262;
+    margin-bottom: 10px;
+  }
+}
+</style>
