+++ pages/create-quest.vue
@@ -61,87 +61,7 @@
               />
             </div>
           </div>
-          <div class="page__skills skills">
-            <div
-              v-for="key in specCount"
-              :key="key"
-              class="skills__block block"
-            >
-              <div class="block__skill-spec">
-                <div class="block__specialization specialization">
-                  <base-dd
-                    v-model="displaySpecIndex[key]"
-                    class="specialization__dd"
-                    type="gray"
-                    :items="specializationsNames"
-                    :placeholder="$t('settings.selectSpec')"
-                    :mode="'small'"
-                    rules="required"
-                    :label="$t('settings.specialization')"
-                    :hide-selected="hideSelectedSpecs"
-                    @input="switchSkill($event, key)"
-                  />
-                  <div class="specialization__skills skills">
-                    <base-dd
-                      v-model="skillIndex[key]"
-                      class="specialization__dd"
-                      :type="specIndex[key] < 0 ? 'disabled' : 'gray'"
-                      :disabled="specIndex[key] < 0"
-                      :placeholder="$t('settings.selectSkills')"
-                      :items="skillsNames[displaySpecIndex[key]]"
-                      :mode="'small'"
-                      rules="required"
-                      :label="$t('settings.skillsInput')"
-                      :hide-selected="hideSelectedSkills[key]"
-                      @input="addSkillToBadge($event, key)"
-                    />
-                    <div
-                      v-if="selectedSkills[key].length === 5"
-                      class="skills__error"
-                    >
-                      {{ $t('ui.buttons.errors.manySkills') }}
-                    </div>
-                  </div>
-                </div>
-                <div class="block__skill skill">
-                  <div
-                    v-for="(item, i) in selectedSkills[key]"
-                    :key="i"
-                    class="skill__badge"
-                  >
-                    {{ item.name }}
-                    <button
-                      class="skill__remove"
-                      @click="removeSkillToBadge(item, key)"
-                    >
-                      <img
-                        src="~assets/img/ui/close_blue.svg"
-                        alt="x"
-                      >
-                    </button>
-                  </div>
-                </div>
-              </div>
-              <base-btn
-                :text="$t('settings.removeSpec')"
-                class="specialization__btn specialization__btn_remove"
-                @click="removeSpecialization(key)"
-              />
-            </div>
-            <base-btn
-              :text="$t('settings.addSpec')"
-              :disabled="specCount === 3"
-              class="skills__btn-add"
-              :class="specCount === 3 ? 'skills__btn-add_disabled' : ''"
-              @click="addSpecialization"
-            />
-            <div
-              v-if="specCount === 3"
-              class="skills__error"
-            >
-              {{ $t('ui.buttons.errors.manySpec') }}
-            </div>
-          </div>
+          <specializations-selector @changeSkills="updateSelectedSkills" />
           <div class="page__address">
             <base-field
               v-model="address"
@@ -206,7 +126,7 @@
           <div class="upload btn btn__container btn__container_right">
             <div class="btn__create">
               <base-btn
-                :disabled="!(invalid === false && !(selectedSkills[1].length === 0))"
+                :disabled="!(invalid === false && !(selectedSpecAndSkills.length === 0))"
                 @click="handleSubmit(toRiseViews)"
               >
                 {{ $t('quests.createAQuest') }}
@@ -302,7 +222,7 @@
               <div class="btn-container__btn">
                 <base-btn
                   :mode="'outline'"
-                  @click="toCreateQuest()"
+                  @click="createQuest"
                 >
                   {{ $t('meta.skipAndEnd') }}
                 </base-btn>
@@ -344,12 +264,7 @@ export default {
         currentAdPrice: '',
       },
       period: 1,
-      specCount: 0,
-      specIndex: { 1: -1, 2: -1, 3: -1 }, // Выбранные специализации по id в filters
-      displaySpecIndex: { 1: -1, 2: -1, 3: -1 }, // Выбранные для отображения названия специализации
-      hideSelectedSkills: { 1: [], 2: [], 3: [] },
-      skillIndex: { 1: -1, 2: -1, 3: -1 },
-      selectedSkills: { 1: [], 2: [], 3: [] },
+      selectedSpecAndSkills: [],
       employmentIndex: 0,
       workplaceIndex: 0,
       runtimeIndex: 0,
@@ -508,59 +423,15 @@ export default {
         this.$t('quests.distantWork.bothVariant'),
       ];
     },
-    specializationsNames() {
-      const specs = [];
-      const keys = Object.keys(this.filters);
-      for (let i = 0; i < keys.length; i += 1) {
-        const key = keys[i];
-        specs.push(this.$t(`filters.items.${this.filters[key].id}.title`));
-      }
-      return specs;
-    },
-    specsIndexes() { // id спеки из filters
-      const specsIndexes = [];
-      const keys = Object.keys(this.filters);
-      for (let i = 0; i < keys.length; i += 1) {
-        const key = keys[i];
-        specsIndexes.push(this.filters[key].id);
-      }
-      return specsIndexes;
-    },
-    skillsNames() {
-      const skillsData = {};
-      const keys = Object.keys(this.filters);
-      for (let i = 0; i < keys.length; i += 1) {
-        const key = keys[i];
-        const skillsBySpec = this.filters[key].skills;
-        const skillsIndexes = Object.values(skillsBySpec);
-        const skillsNames = [];
-        for (let j = 0; j < skillsIndexes.length; j += 1) {
-          const name = this.$t(`filters.items.${this.filters[key].id}.sub.${skillsIndexes[j]}`);
-          skillsNames.push(name);
-        }
-        skillsData[i] = skillsNames;
-      }
-      return skillsData;
-    },
-    skillsIndexes() {
-      const skillsIndexesData = {};
-      const keys = Object.keys(this.filters);
-      for (let i = 0; i < keys.length; i += 1) {
-        const key = keys[i];
-        const skillsBySpec = this.filters[key].skills;
-        skillsIndexesData[i] = Object.values((skillsBySpec)); // skills indexes from filters
-      }
-      return skillsIndexesData;
-    },
-    hideSelectedSpecs() {
-      return Object.values(this.displaySpecIndex);
-    },
   },
   async mounted() {
     this.SetLoader(true);
     this.SetLoader(false);
   },
   methods: {
+    updateSelectedSkills(specAndSkills) {
+      this.selectedSpecAndSkills = specAndSkills;
+    },
     cardStatus(item) {
       let style;
       if (item.code === 1) {
@@ -617,53 +488,6 @@ export default {
     goBack() {
       this.$store.dispatch('quests/getCurrentStepCreateQuest', 1);
     },
-    addSkillToBadge(event, key) {
-      const index = this.skillsIndexes[this.displaySpecIndex[key]][event];
-      if (this.selectedSkills[key].length > 4
-        || this.selectedSkills[key].filter((obj) => obj.index === index).length) return;
-      this.selectedSkills[key].push({
-        index,
-        name: this.skillsNames[this.displaySpecIndex[key]][event],
-      });
-      this.hideSelectedSkills[key].push(event);
-    },
-    removeSkillToBadge(skill, key) {
-      let hideIndex = null;
-      for (let i = 0; i < this.skillsIndexes[key].length; i += 1) {
-        if (this.skillsIndexes[this.displaySpecIndex[key]][i] === skill.index) {
-          hideIndex = i;
-          break;
-        }
-      }
-      const numberInHide = this.hideSelectedSkills[key].indexOf(hideIndex);
-      if (numberInHide > -1) {
-        this.hideSelectedSkills[key].splice(numberInHide, 1);
-      }
-
-      const numberInArray = this.selectedSkills[key].indexOf(skill);
-      this.selectedSkills[key].splice(numberInArray, 1);
-      if (!this.selectedSkills[key].length) {
-        this.skillIndex[key] = -1;
-      }
-    },
-    switchSkill(event, key) {
-      this.specIndex[key] = this.specsIndexes[event];
-      this.skillIndex[key] = -1;
-      this.selectedSkills[key] = [];
-      this.hideSelectedSkills[key] = [];
-    },
-    addSpecialization() {
-      if (this.specCount <= 2) {
-        this.specCount += 1;
-      }
-    },
-    removeSpecialization(key) {
-      this.selectedSkills[key] = [];
-      this.specIndex[key] = -1;
-      this.displaySpecIndex[key] = -1;
-      this.skillIndex[key] = -1;
-      this.specCount -= 1;
-    },
     selectAddress(address) {
       this.addresses = [];
       this.address = address.formatted;
@@ -697,21 +521,7 @@ export default {
         console.log(e);
       }
     },
-    async toCreateQuest() {
-      const specAndSkills = [];
-      let key = 1;
-      // eslint-disable-next-line guard-for-in,no-restricted-syntax
-      for (const spec in this.specIndex) {
-        if (this.specIndex[spec] !== -1) {
-          for (let i = 0; i < this.selectedSkills[key].length; i += 1) {
-            specAndSkills.push(`${this.specIndex[spec]}.${this.selectedSkills[key][i].index}`);
-          }
-        }
-        key += 1;
-      }
-      await this.createQuest(specAndSkills);
-    },
-    async createQuest(specAndSkills) {
+    async createQuest() {
       const payload = {
         workplace: this.convertWorkplace(this.workplaceIndex),
         priority: this.runtimeIndex,
@@ -723,7 +533,7 @@ export default {
         medias: [],
         adType: 0,
         locationPlaceName: this.address,
-        specializationKeys: specAndSkills,
+        specializationKeys: this.selectedSpecAndSkills,
         location: {
           longitude: this.coordinates.lng,
           latitude: this.coordinates.lat,
@@ -1167,55 +977,6 @@ export default {
 }
 
 .page {
-  &__skills {
-    width: 100%;
-    .block {
-      display: flex;
-      grid-gap: 20px;
-      justify-content: space-between;
-      margin-top: 20px;
-      &__skill-spec {
-        width: 100%;
-      }
-      &__specialization {
-        display: flex;
-        align-items: flex-start;
-        flex-direction: row;
-        grid-gap: 20px;
-      }
-      &__skill {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        flex-wrap: wrap;
-        grid-gap: 10px;
-        .skill {
-          &__badge {
-            background: rgba(0, 131, 199, 0.1);
-            border-radius: 44px;
-            color: $blue;
-            white-space: nowrap;
-            grid-gap: 8px;
-            padding: 5px 10px 5px 10px;
-            display: flex;
-            text-align: center;
-            &-skills {
-              padding: 15px;
-            }
-          }
-        }
-      }
-    }
-  }
-  &__spec {
-    flex-direction: row;
-    flex-wrap: wrap;
-    display: flex;
-    max-width: 1180px;
-    width: 100%;
-    justify-content: flex-start;
-    //padding: 0 20px 0 0;
-  }
   &__title {
     @include text-simple;
     margin: 30px 0 0 0;
@@ -1261,57 +1022,7 @@ export default {
     }
   }
 }
-.specialization {
-  &__dd {
-    margin-bottom: 15px;
-    width: 100%;
-  }
-  &__skills {
-    width: 100%;
-  }
-  &__btn {
-    text-align: center;
-    &_remove {
-      margin-top: 37px;
-      width: 50%;
-      background: #ffffff;
-      color: #d73838;
-      border: 1px solid #e79a9a;
-    }
-    &_remove:hover {
-      background: #e79a9a;
-    }
-  }
-}
-.skills {
-  &__add-info {
-    display: flex;
-    flex-wrap: nowrap;
-    grid-gap: 20px;
-    margin-top: 20px;
-  }
-  &__btn {
-    &-add {
-      text-align: center;
-      margin-top: 20px;
-      width: 250px;
-      background: #FFFFFF;
-      color: #0083C7;
-      border: 1px solid #bce8ff;
-      &_disabled {
-        color: #aaaaaa !important;
-        border: 0.5px solid #aaaaaa;
-      }
-    }
-    &-add:hover {
-      background: #bce8ff;
-    }
-  }
-  &__error {
-    color: #f36262;
-    margin-bottom: 10px;
-  }
-}
+
 .main {
   @include main;
   &-white {
