+++ plugins/injectComponents.js
@@ -21,6 +21,7 @@ import QuestIdDD from '~/components/ui/QuestIdDD';
 import InvitedWorkerList from '~/components/app/pages/quests_id/invitedWorkerList';
 import RespondedWorkerList from '~/components/app/pages/quests_id/respondedWorkerList';
 import StarRating from '~/components/app/pages/quests_id/starRating';
+import SpecializationsSelector from '~/components/ui/SpecializationsSelector';
 
 Vue.component('ctm-modal', CtmModal);
 Vue.component('ctm-modal-box', CtmModalBox);
@@ -43,3 +44,4 @@ Vue.component('vue-phone-number-input', VuePhoneNumberInput);
 Vue.component('invited-worker-list', InvitedWorkerList);
 Vue.component('responded-worker-list', RespondedWorkerList);
 Vue.component('star-rating', StarRating);
+Vue.component('specializations-selector', SpecializationsSelector);
